fix!: Increase u16 fields to u32
Failed to deserialize iccprofile because value was too big. Switch to u32 for other u16 fields as well, in case.

+++ mediaflow_core/src/folders.rs
@@ -20,7 +20,7 @@ pub struct FolderFull {
     pub name: String,
     #[serde(rename = "type")]
     pub type_: u32,
-    pub depth: u16,
+    pub depth: u32,
     pub parent: Option<u32>,
     pub created: String,
     pub created_by: u32,
@@ -48,7 +48,7 @@ pub struct ParentFolder {
     pub name: String,
     #[serde(rename = "type")]
     pub type_: u32,
-    pub depth: u16,
+    pub depth: u32,
     pub parent: Option<u32>,
     pub has_children: bool,
 }
